Revert "fixed models and relationships"
This reverts commit 35355a77fce824b8cf70ee8cd2297b41bdbdb573.
