+++ src/server/app/main/models/ChallengesModel.py
@@ -16,9 +16,6 @@ class ChallengesModel(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     challenge_name = db.Column(db.String(80), nullable=False)
     description = db.Column(db.Text, nullable=False)
-    problem_statement = db.Column(db.Text, nullable=False)
-    input_format = db.Column(db.Text, nullable=False)
-    output_format = db.Column(db.Text, nullable=False)
     difficulty = db.Column(db.String(20), nullable=False)
     sample_input = db.Column(db.Text, nullable=False)
     sample_output = db.Column(db.Text, nullable=False)
@@ -26,7 +23,6 @@ class ChallengesModel(db.Model):
                            nullable=False, default=datetime.datetime.now())
     challenge_settings = db.relationship(
         'ChallengeSettings', backref='challenge', lazy=True)
-    tags = db.relationship('TagsModel', secondary=ChallengesTagsModel.challenges_tags, lazy='subquery',
+    tags = db.relationship('Tags', secondary=ChallengesTagsModel.challenges_tags, lazy='subquery',
                            backref=db.backref('challenges', lazy=True))
-    test_cases = db.relationship(
-        'TestCasesModel', backref='challenge', lazy=True)
+    test_cases = db.relationship('TestCases', backref='challenge', lazy=True)
