+++ src/server/app/main/models/ContestsModel.py
@@ -1,9 +1,7 @@
 from .. import db
 import datetime
-# from . import UsersContestsModel
-# from . import ContestsChallengesModel
-from .ContestsChallengesModel import contests_challenges
-from .UsersContestsModel import users_contests
+from . import UsersContestsModel
+from . import ContestsChallengesModel
 
 
 class ContestsModel(db.Model):
@@ -24,20 +22,7 @@ class ContestsModel(db.Model):
     show_leaderboard = db.Column(db.Boolean, nullable=False, default=True)
     created_at = db.Column(db.DateTime(timezone=False),
                            nullable=False, default=datetime.datetime.now())
-    users_contests = db.relationship('UserModel', secondary=users_contests, lazy='subquery',
+    users_contests = db.relationship('Users', secondary=UsersContestsModel.users_contests, lazy='subquery',
                                      backref=db.backref('contests', lazy=True))
-    contests_challenges = db.relationship('ChallengesModel', secondary=contests_challenges, lazy='subquery',
+    contests_challenges = db.relationship('Challenges', secondary=ContestsChallengesModel.contests_challenges, lazy='subquery',
                                           backref=db.backref('contests', lazy=True))
-
-    @classmethod
-    def find_by_name(cls, contest_name):
-        """
-            Find by contest Name
-            :param contest_name
-            :return object containing all contest details
-        """
-        return cls.query.filter_by(contest_name=contest_name)
-
-    @classmethod
-    def get_contests_challenges(cls, contest_name):
-        return cls.query.filter_by(contest_name=contest_name).all()
