+++ src/server/app/main/models/UsersModel.py
@@ -1,8 +1,5 @@
 from .. import db
 import datetime
-from app.main import flask_bcrypt, db
-import jwt
-from app.main.settings import key
 
 
 class UserModel(db.Model):
@@ -16,9 +13,7 @@ class UserModel(db.Model):
     __tablename__ = "users"
 
     id = db.Column(db.Integer, primary_key=True)
-    public_id = db.Column(db.String(50), unique=True)
     name = db.Column(db.String(80), nullable=False)
-    password_hash = db.Column(db.String(255), nullable=False)
     email = db.Column(db.String(120), unique=True, nullable=False)
     role = db.Column(db.String(20), nullable=False, default="user")
     profile_location = db.Column(db.String(250), nullable=True)
@@ -31,57 +26,3 @@ class UserModel(db.Model):
         db.session.add(cls(
             **{"name": "sachin2", "email": "beckhamd@gmail.com", "role": "admin", "salt": "xyz"}))
         db.session.commit()
-
-    @property
-    def password(self):
-        raise AttributeError('password: write-only field')
-
-    @password.setter
-    def password(self, password):
-        self.password_hash = flask_bcrypt.generate_password_hash(
-            password).decode('utf-8') if password else None
-
-    def check_password(self, password):
-        return flask_bcrypt.check_password_hash(self.password_hash, password)
-
-    @staticmethod
-    def encode_auth_token(user_id):
-        """
-        Generates the Auth Token
-        :return: string
-        """
-        try:
-            payload = {
-                'exp': datetime.datetime.utcnow() + datetime.timedelta(days=1, seconds=5),
-                'iat': datetime.datetime.utcnow(),
-                'sub': user_id
-            }
-            return jwt.encode(
-                payload,
-                key,
-                algorithm='HS256'
-            )
-        except Exception as e:
-            return e
-
-    @staticmethod
-    def decode_auth_token(auth_token):
-        """
-        Decodes the auth token
-        :param auth_token:
-        :return: integer|string
-        """
-        try:
-            payload = jwt.decode(auth_token, key)
-            # is_blacklisted_token = BlacklistToken.check_blacklist(auth_token)
-            # if is_blacklisted_token:
-            #     return 'Token blacklisted. Please log in again.'
-            # else:
-            #     return payload['sub']
-        except jwt.ExpiredSignatureError:
-            return 'Signature expired. Please log in again.'
-        except jwt.InvalidTokenError:
-            return 'Invalid token. Please log in again.'
-
-    def __repr__(self):
-        return "<User '{}'>".format(self.username)
