+++ src/server/app/main/routes/__init__.py
@@ -1,12 +1,10 @@
 from app.main.routes.auth_controller import UserLogin, LogoutAPI, UserSignUp, FacebookAuthorize, GithubAuthorize
 from app.main import api
-from app.main.routes.Contest import Contest
-
 
 def add_resources(app):
     """
     Method to add resources to app context
-
+    
     Args:
         app (object): object of Flask representing the app in context
     """
@@ -15,13 +13,11 @@ def add_resources(app):
     api.add_resource(UserSignUp, '/signup')
     api.add_resource(FacebookAuthorize, '/facebook')
     api.add_resource(GithubAuthorize, '/github')
-    api.add_resource(Contest, '/contest/<contest_name>')
-
 
 def register_blueprints(app):
     """
     Method to add blueprints to app context
-
+    
     Args:
         app (object): object of Flask representing the app in context
     """
