+++ src/server/app/main/routes/blueprint_test.py
@@ -2,7 +2,6 @@ from flask import Blueprint, current_app
 from ..models.UsersModel import UserModel
 bp = Blueprint("public", __name__)
 
-
 @bp.route("/")
 def hello_world():
     current_app.logger.info("Here I am")
