+++ src/server/app/main/services/auth_helper.py
@@ -1,4 +1,4 @@
-from app.main.models.UsersModel import UserModel
+from app.main.models.user import User
 
 
 class Auth:
@@ -7,7 +7,7 @@ class Auth:
         # import pdb; pdb.set_trace()
         try:
             # fetch the user data
-            user = UserModel.query.filter_by(email=data.get('email')).first()
+            user = User.query.filter_by(email=data.get('email')).first()
             if user and user.check_password(data.get('password')):
                 auth_token = user.encode_auth_token(user.id)
                 if auth_token:
@@ -39,7 +39,7 @@ class Auth:
         else:
             auth_token = ''
         if auth_token:
-            resp = UserModel.decode_auth_token(auth_token)
+            resp = User.decode_auth_token(auth_token)
             if not isinstance(resp, str):
                 # mark the token as blacklisted
                 # return save_token(token=auth_token)
@@ -56,3 +56,4 @@ class Auth:
                 'message': 'Provide a valid auth token.'
             }
             return response_object, 403
+
